FIX: Now non-data files are removed from the fn_list from the ftp server
It seems that the DWD added a pdf file to the data dir on 2017-07-21 which
caused the current code to stop since it tried to derive a timestamp from
the first file in the fn_list from the ftp, which now was the pdf, which
obviously had not timestamp encoded in the file name. Now only files that
have the same ending as the RADOLAN files are kept in the update
fn_list.