Add and use NetCDF with RADOLAN lat-lon coordinates
This was necessary since newer versions of wradlib and the newer
versions of the dependecies seem to produce slightly different
latitude values (somehow random differences of approx 1e-14 degree).
By using this fixed set ot coordinates this problem can be ruled out
for the future.

The provided cooridnates are the ones that were produce with wradlib
at some point in 2018.